Combine provider / controls hooks
