+++ apps/react/demo/src/ui/Example/Provider/Provider.tsx
@@ -1,9 +1,14 @@
 import Context from "../Context.js";
+import { useProviderState } from "../hooks/index.js";
 import type { ProviderProps } from "./types.js";
-import useProviderState from "./useProviderState.js";
 
-const Provider = ({ items = [], defaultValues, children }: ProviderProps) => {
-  const state = useProviderState({ items, defaultValues });
+const Provider = ({
+  items,
+  defaultValues,
+  children,
+  outputFormats,
+}: ProviderProps) => {
+  const state = useProviderState({ items, defaultValues, outputFormats });
 
   return <Context.Provider value={state}>{children}</Context.Provider>;
 };
