--- apps/react/demo/src/ui/Example/hooks/useExampleControls/index.ts
@@ -1,2 +1,0 @@
-export { default as useExampleControls } from "./useExampleControls.js";
-export * from "./types.js";
