--- apps/react/demo/src/ui/Example/hooks/useExampleControls/types.ts
@@ -1,5 +1,0 @@
-export type UseExampleControlsResult = {
-  handleCopyCss: () => void;
-  handlePrevExample: () => void;
-  handleNextExample: () => void;
-};
